fix: int32 -> int64

+++ services/billing-management-service/internal/repository/analytics.go
@@ -165,7 +165,7 @@ func (r *AnalyticsRepository) GetRevenueMetrics(ctx context.Context) (*models.Re
 		metrics.MonthlyRevenue = result["monthlyRevenue"].(float64)
 		metrics.YearlyRevenue = result["yearlyRevenue"].(float64)
 		metrics.AveragePrice = result["avgPrice"].(float64)
-		subscriberCount := result["subscriberCount"].(int32)
+		subscriberCount := result["subscriberCount"].(int64)
 
 		if subscriberCount > 0 {
 			metrics.AverageRevenuePerUser = metrics.MonthlyRevenue / float64(subscriberCount)
@@ -241,7 +241,7 @@ func (r *AnalyticsRepository) GetSubscriptionTrends(ctx context.Context, period 
 	for i, result := range results {
 		trends.Data[i] = models.TrendData{
 			Period: result["_id"],
-			Count:  result["count"].(int32),
+			Count:  result["count"].(int64),
 		}
 	}
